Rename constants.js > config.js
