+++ src/config.js
@@ -1,7 +1,6 @@
-/* DEFINE CONSTANTS */
-
 const ANIMATION_SLIDE_DURATION = 400; // ms
 const ARROW_PRESS_TIMEOUT = ANIMATION_SLIDE_DURATION;
+
 
 module.exports = {
   ARROW_PRESS_TIMEOUT,
